Silence signed/unsigned comparison warning for file size
off_t is signed, size_t is not, but the size of a struct cache_file
definitely fits in an off_t.

+++ utils/ld-cache.c
@@ -123,7 +123,7 @@ ld_cache_open (ldcache_t *cache, const char *path)
     fstat( cache->fd, &ldcache );
 
     // cache file must be at least this big or it's invalid:
-    if( ldcache.st_size < sizeof( struct cache_file ) )
+    if( ldcache.st_size < (off_t) sizeof( struct cache_file ) )
         goto cleanup;
 
     cache->mmap =
